tests/wpt/web-platform-tests/html/semantics/text-level-semantics/the-a-element/a.rel-getter-01.html

<!DOCTYPE html>
<title>HTMLAnchorElement.rel getter</title>
<link rel="author" title="TheKK" href="mailto:thumbd03803@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-a-rel">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<div>
<a id="test" href="a" rel="noreferrer"></a>
</div>
<script>
test(function() {
  var a = document.getElementById("test");

  test(function() {
    assert_equals(a.rel, "noreferrer");
  }, "Test anchor's rel getter");
});
</script>
